requirements_venv.txt
absl-py==2.3.1
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
aiosqlite==0.21.0
airportsdata==20250811
annotated-types==0.7.0
anthropic==0.64.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
# Agent Frameworks (compatible with Python 3.13)
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext[docker]==0.7.5
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.10
langgraph==0.6.11
llama-index==0.12.31
llama-index-llms-openai==0.3.28
llama-index-core==0.12.31
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
banks==2.2.0
beartype==0.12.0
beautifulsoup4==4.13.4
blake3==1.0.5
blinker==1.9.0
Brotli==1.1.0
browsergym==0.14.2
browsergym-assistantbench==0.14.2
browsergym-core==0.14.2
browsergym-experiments==0.14.2
browsergym-miniwob==0.14.2
browsergym-visualwebarena==0.14.2
browsergym-webarena==0.14.2
browsergym-workarena==0.4.4
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorama==0.4.6
compressed-tensors==0.10.2
contourpy==1.3.3
cupy-cuda12x==13.5.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
depyf==0.19.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dm-tree==0.1.9
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
e2b==1.11.1
e2b-code-interpreter==1.5.2
einops==0.8.1
email_validator==2.2.0
english-words==2.0.2
eval_type_backport==0.2.2
evaluate==0.4.5
executing==2.2.0
faiss-cpu==1.12.0
Faker==37.5.3
Farama-Notifications==0.0.4
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastcore==1.8.7
fastrlock==0.8.3
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
firecrawl==2.16.5
Flask==3.1.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
genson==1.3.0
gguf==0.17.1
ghapi==1.0.6
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.42.0
gradio_client==1.11.1
greenlet==3.0.3
griffe==1.12.1
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.62.3
grpclib==0.4.8
gym==0.26.2
gym-notices==0.1.0
gym-sokoban==0.0.6
gymnasium==1.2.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hydra-core==1.3.2
math_verify
hyperframe==6.1.0
identify==2.6.13
idna==3.10
imageio==2.37.0
importlib_metadata==8.0.0
interegular==0.3.3
ipython==9.4.0
ipython_pygments_lexers==1.1.1
iso3166==2.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpath-ng==1.7.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kubernetes==33.1.0
lark==1.2.2
latex2sympy2==1.5.4
lazy_loader==0.4
libvisualwebarena==0.0.15
libwebarena==0.0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.10.1
mdurl==0.1.2
mistral_common==1.8.3
modal==1.1.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.55
oauthlib==3.3.1
omegaconf==2.3.0
openai==1.90.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.2
outcome==1.3.0.post0
outlines
outlines_core
packaging==25.0
pandas==2.3.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.17.0
vllm==0.10.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.44.0
ply==3.11
polars==1.32.3
portalocker==3.2.0
pre_commit==4.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pybind11==3.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==11.1.0
pygame==2.6.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.3
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
ray==2.48.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.9
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
sentry-sdk==2.35.0
setproctitle==1.3.6
setuptools==79.0.1
sglang==0.4.10.post2
shapely==2.1.1
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
swebench==4.0.4
sympy==1.14.0
synchronicity==0.10.2
tabulate==0.9.0
tenacity==9.1.2
tensordict==0.8.3
termcolor==3.1.0
text-generation==0.7.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
together==1.5.25
tokenizers==0.21.4
toml==0.10.0
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.4
types-certifi==2021.10.8.3
types-requests==2.32.4.20250809
types-toml==0.10.8.20240310
types-tqdm==4.67.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
unidiff==0.7.5
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vertexai==1.71.1
virtualenv==20.34.0
wandb==0.21.1
watchfiles==1.1.0
wcwidth==0.2.13
weblinx==0.3.2
weblinx_browsergym==0.0.3
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
